pointerevents/pointerevent_pointerLeave_touch.html

<!doctype html>
<html>
    <head>
        <title>Pointer Event: Dispatch pointerLeave (touch). </title>
        <meta content="text/html; charset=UTF-8" http-equiv="Content-Type"/>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/"/>
        <meta name="assert" content="When a pointing device that does not support hover (such as a finger) leaves the hit test boundaries as a result of a pointerup event, the pointerleave event must be dispatched. "/>
        
        <!-- http://w3c-test.org/resources/testharness.js -->
	    <script src="http://w3c-test.org/resources/testharness.js"></script>
		<!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>
        <script type="text/javascript">

            setup({ explicit_done: true });

            var detected_pointertypes = {};
            add_completion_callback(showPointerTypes);

            var target0;
            var test_pointerLeave;

            function run() {
                target0 = document.getElementById("target0");
                test_pointerLeave = async_test("pointerleave event");  // set up test harness

                on_event(target0, "pointerleave", function (event) {
                    check_PointerEvent(event);
                    test_pointerLeave.step(function () {
                        assert_equals(event.pointerType, "touch", " Test should be run using touch input");
                        assert_equals(event.type, "pointerleave", "The " + event.type + " event received");
                    });
                    detected_pointertypes[event.pointerType] = true;
                    test_pointerLeave.done(); // complete test
                });
            }

            function showPointerTypes() {
                var complete_notice = document.getElementById("complete-notice");
                var pointertype_log = document.getElementById("pointertype-log");
                var pointertypes = Object.keys(detected_pointertypes); pointertype_log.innerHTML = pointertypes.length ? pointertypes.join(",") : "(none)";
                if (test_pointerLeave.status == test_pointerLeave.NOTRUN
                    || test_pointerLeave.status == test_pointerLeave.TIMEOUT) {
                    complete_notice.innerHTML = "Test FAIL due to time out.  This is due to the browser not supporting pointer, or the manual steps were not run to completion.";
                    complete_notice.style.background = "red";
                }
                complete_notice.style.display = "block";
            }
        </script>

            <style>
            div {
                margin: 0em;
                padding: 2em;
            }

            #target0 {
                background: yellow;
                border: 1px solid orange;
            }

            #complete-notice {
                border: 1px solid #0a0;
                display: none;
            }

            #pointertype-log {
                font-weight: bold;
            }
        </style>
    </head>
    <body onload="run()">
        <h1>Pointer Event: Dispatch pointerleave (touch)</h1>
        <h4> 
            Test Description: 
            When a pointing device that does not support hover (such as a finger) leaves the hit test boundaries as a result of a pointerup event, the pointerleave event must be dispatched.  
        </h4>
        <br />
        <div id="target0">
            Use touch to tap on this box.
        </div>
        <div id="complete-notice">
            <p>Test complete:  Scroll to Summary to view Pass/Fail Results.</p>
            <p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
        </div>

        <div id="log"></div>
    </body>
</html>
